requirements.txt
yamlu==0.0.17
tqdm==4.66.4
torchvision==0.18.0
azure-ai-vision-imageanalysis==1.0.0b2
gdown
psutil
fastapi
uvicorn
gradio 
pillow
